Move cca cli codes under pkg/cli/
