+++ cmd/cca/cca.go
@@ -22,9 +22,9 @@ import (
 	"github.com/cloud-ca/cca/cmd/cca/connection"
 	"github.com/cloud-ca/cca/cmd/cca/version"
 	"github.com/cloud-ca/cca/pkg/cli"
-	"github.com/cloud-ca/cca/pkg/client"
-	"github.com/cloud-ca/cca/pkg/flags"
-	"github.com/cloud-ca/cca/pkg/output"
+	"github.com/cloud-ca/cca/pkg/cli/client"
+	"github.com/cloud-ca/cca/pkg/cli/flags"
+	"github.com/cloud-ca/cca/pkg/cli/output"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
