+++ cmd/cca/connection/connection.go
@@ -19,7 +19,7 @@ import (
 	"github.com/cloud-ca/cca/cmd/cca/connection/get"
 	"github.com/cloud-ca/cca/cmd/cca/connection/list"
 	"github.com/cloud-ca/cca/pkg/cli"
-	"github.com/cloud-ca/cca/pkg/util"
+	"github.com/cloud-ca/cca/pkg/cli/util"
 	"github.com/spf13/cobra"
 )
